proxy_requirements.txt
aiohttp>=3.9
pyyaml>=5.0
